+++ src/test/crimson/CMakeLists.txt
@@ -17,7 +17,11 @@ target_link_libraries(perf_crimson_msgr ceph-common crimson)
 
 add_executable(unittest_seastar_echo
   test_alien_echo.cc)
-target_link_libraries(unittest_seastar_echo ceph-common global crimson)
+target_link_libraries(unittest_seastar_echo crimson)
+
+add_executable(unittest_async_echo
+  test_async_echo.cc)
+target_link_libraries(unittest_async_echo ceph-common global)
 
 add_executable(unittest_seastar_thread_pool
   test_thread_pool.cc)
